requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
cobble==0.1.3
Django==3.1.3
django-bootstrap4==2.3.1
django-parler==2.2
django-polymorphic==3.0.0
django-recaptcha==2.0.6
django-tag-parser==3.2
django-template-analyzer==1.6.2
docxtpl==0.11.2
future==0.18.2
html5lib==1.1
idna==2.10
Jinja2==2.11.2
lxml==4.6.2
mammoth==1.4.13
MarkupSafe==1.1.1
mysqlclient==2.0.1
Pillow==8.0.1
python-docx==0.8.10
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tqdm==4.54.0
urllib3==1.26.2
webencodings==0.5.1
